+++ application/models/Flight.php
@@ -1,89 +1,161 @@
 <?php
-// First letter of the airline, Yellow Hammer.
-define("IDENTIFYING_LETTER", 'Y');
+
+require_once(APPPATH . 'core/Entity.php');
+
 /**
- * Flights model. Contains the list of scheduled flights.
- *
- * @author Ian Park
+ * Defines a flight.
  */
-class Flight extends CI_Model
-{
-	
-	// const IDENTIFYING_LETTER = 'Y';
-	var $data = [
-		0 => [
-			'id' => IDENTIFYING_LETTER . '001',
-			'departure_airport' => 0,
-			'departure_time' => '10:30',
-			'arrival_airport' => 1,
-			'arrival_time' => '19:10',
-			'day' => 'Monday'
-		],
-		1 => [
-			'id' => IDENTIFYING_LETTER . '002',
-			'departure_airport' => 0,
-			'departure_time' => '11:30',
-			'arrival_airport' => 2,
-			'arrival_time' => '15:15',
-			'day' => 'Tuesday'
-		],
-		2 => [
-			'id' => IDENTIFYING_LETTER . '003',
-			'departure_airport' => 1,
-			'departure_time' => '18:30',
-			'arrival_airport' => 2,
-			'arrival_time' => '06:40',
-			'day' => 'Monday'
-		],
-		3 => [
-			'id' => IDENTIFYING_LETTER . '004',
-			'departure_airport' => 1,
-			'departure_time' => '09:00',
-			'arrival_airport' => 3,
-			'arrival_time' => '16:50',
-			'day' => 'Friday'
-		],
-		4 => [
-			'id' => IDENTIFYING_LETTER . '005',
-			'departure_airport' => 3,
-			'departure_time' => '12:30',
-			'arrival_airport' => 4,
-			'arrival_time' => '15:55',
-			'day' => 'Monday'
-		]
-	];
-	// Constructor
-	public function __construct()
-	{
-		parent::__construct();
-		// Inject each "record" key into the record itself, for ease of presentation.
-		// Plus, inject airport information according to their keys.
-		foreach ($this->data as $key => $record)
-		{
-			require_once('Airports.php');
-			$record['key'] = $key;
-			$this->data[$key] = $record;
-			$this->data[$key]['departure_airport'] = [(new Airports)->get($record['departure_airport'])];
-			$this->data[$key]['arrival_airport'] = [(new Airports)->get($record['arrival_airport'])];
+class Flight extends Entity {
+
+	private $id;
+	private $departure_airport;
+	private $departure_time;
+	private $arrival_airport;
+	private $arrival_time;
+	private $day;
+
+	/**
+	 * Sets the flight ID.
+	 * 
+	 * @param string    $value    ID of the flight
+	 */
+	public function setId($value) {
+		// alpha_numeric|greater_than[0]|less_than[1000]
+		if (preg_match('/^Y[1-9][0-9]{1,2}$/i', $value)) {
+			$letter = strtoupper(substr($value, 0, 1));
+			$number = intval(substr($value, 1));
+
+			if ($number != 0) {
+				$number = sprintf('%03d', $value);
+			}
+
+			$this->id = $letter . $number;
 		}
+
+		return;
 	}
+
 	/**
-	 * Retrieves a single flight schedule, null if not found.
+	 * Gets the flight ID.
 	 * 
-	 * @param  int 			$which 	the key of a flgit schedule
-	 * @return array|null        	a flight schedule, or null if none
+	 * @return string    ID of the flight
 	 */
-	public function get($which)
-	{
-		return !isset($this->data[$which]) ? null : $this->data[$which];
+	public function getId() {
+		return $this->id;
 	}
+
 	/**
-	 * Retrieves all the schedules.
+	 * Sets the departure airport.
 	 * 
-	 * @return array the flight schedules
+	 * @param string    $value    ID of the departure airport
 	 */
-	public function all()
-	{
-		return $this->data;
+	public function setDepartureAirport($value) {
+		// alpha|exact_length[3]
+		if (preg_match('/^[a-zA-Z]{3}$/', $value)) {
+			$this->departure_airport = strtoupper($value);
+		}
+
+		return;
+	}
+
+	/**
+	 * Gets the departure airport.
+	 * 
+	 * @return string    ID of the departure airport
+	 */
+	public function getDepartureAirport() {
+		return $this->departure_airport;
+	}
+
+	/**
+	 * Sets the departure time.
+	 * 
+	 * @param string    $value    Departure time
+	 */
+	public function setDepartureTime($value) {
+		// regex_match[^(2[0-3]|[01][0-9]):([0-5][0-9])$]
+		if (preg_match('/^(2[0-3]|[01][0-9]):([0-5][0-9])$/', $value)) {
+			$this->departure_time = $value;
+		}
+
+		return;
+	}
+
+	/**
+	 * Gets the departure time.
+	 * 
+	 * @return string    Departure time
+	 */
+	public function getDepartureTime() {
+		return $this->departure_time;
+	}
+
+	/**
+	 * Sets the arrival airport.
+	 * 
+	 * @param string    $value    ID of the arrival airport
+	 */
+	public function setArrivalAirport($value) {
+		// alpha|exact_length[3]
+		if (preg_match('/^[a-zA-Z]{3}$/', $value)) {
+			$this->arrival_airport = strtoupper($value);
+		}
+
+		return;
+	}
+
+	/**
+	 * Gets the region of the airport.
+	 * 
+	 * @return string    ID of the arrival airport
+	 */
+	public function getArrivalAirport() {
+		return $this->arrival_airport;
+	}
+
+	/**
+	 * Sets the departure time.
+	 * 
+	 * @param string    $value    Arrival time
+	 */
+	public function setArrivalTime($value) {
+		// regex_match[^(2[0-3]|[01][0-9]):([0-5][0-9])$]
+		if (preg_match('/^(2[0-3]|[01][0-9]):([0-5][0-9])$/', $value)) {
+			$this->arrival_time = $value;
+		}
+
+		return;
+	}
+
+	/**
+	 * Gets the arrival time.
+	 * 
+	 * @return string    Arrival time
+	 */
+	public function getArrivalTime() {
+		return $this->arrival_time;
+	}
+
+	/**
+	 * Sets the day of week of the flight.
+	 * 
+	 * @param string    $value    Day of week of the flight
+	 */
+	public function setDayOfWeek($value) {
+		// alpha|max_length[9]
+		if (preg_match('/^monday\b|tuesday\b|wednesday\b|thursday\b|friday\b|saturday\b|sunday\b$/i', $value)) {
+			$this->day = $value;
+		}
+
+		return;
+	}
+
+	/**
+	 * Gets the day of week of the flight.
+	 * 
+	 * @return string    Day of week of the flight
+	 */
+	public function getDayOfWeek() {
+		return $this->day;
 	}
 }