+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## 0.0.2 - 2017-11-12
+### Added
+- Model
+  - Entity models (Airplane, Airport, Flight)
+  - Model to retrieve shared data from the WACKY server
+
+### Changed
+- Model
+  - Airplanes, Airports, and Flights models now extend CSV_Model, i.e., gets data from the CSV files
+  - The models have CI rules based on constraints defined in the entity models
+
 ## 0.0.1 - 2017-11-12
 ### Added
 - Other